Redmine 10030: move every method of mobile service into a controller
