+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/module/HumanResourceModule.java
@@ -34,8 +34,6 @@ import com.axelor.apps.hr.db.repo.TimesheetLineHRRepository;
 import com.axelor.apps.hr.db.repo.TimesheetLineRepository;
 import com.axelor.apps.hr.db.repo.TimesheetRepository;
 import com.axelor.apps.hr.db.repo.TimesheetTimerHRRepository;
-import com.axelor.apps.hr.mobile.HumanResourceMobileService;
-import com.axelor.apps.hr.mobile.HumanResourceMobileServiceImpl;
 import com.axelor.apps.hr.service.bankorder.BankOrderServiceHRImpl;
 import com.axelor.apps.hr.service.batch.BatchCreditTransferExpensePaymentHR;
 import com.axelor.apps.hr.service.batch.MailBatchServiceHR;
@@ -90,7 +88,6 @@ public class HumanResourceModule extends AxelorModule {
 		bind(BatchCreditTransferExpensePayment.class).to(BatchCreditTransferExpensePaymentHR.class);
         bind(BankOrderServiceImpl.class).to(BankOrderServiceHRImpl.class);
         bind(HrBatchRepository.class).to(HrBatchHRRepository.class);
-        bind(HumanResourceMobileService.class).to(HumanResourceMobileServiceImpl.class);
 	}
 
 }
